put ProfileChart where it belongs git add -Agit add -A
